Removed the old GetSykmeldingOpplysningerClient
