+++ src/main/kotlin/no/nav/syfo/plugins/KoinDev.kt
@@ -24,9 +24,9 @@ import no.nav.syfo.sykmelding.delete_sykmelding.SykmeldingStatusKafkaProducer
 import no.nav.syfo.sykmelding.delete_sykmelding.SykmeldingStatusKafkaProducerDevelopment
 import no.nav.syfo.sykmelding.delete_sykmelding.TombstoneKafkaProducer
 import no.nav.syfo.sykmelding.delete_sykmelding.TombstoneKafkaProducerDevelopment
-import no.nav.syfo.sykmeldingsopplysninger.DevelopmentSykmeldingsOpplysningerClient
+import no.nav.syfo.sykmeldingsopplysninger.GetSykmeldingOpplysningerDatabase
 import no.nav.syfo.sykmeldingsopplysninger.GetSykmeldingOpplysningerService
-import no.nav.syfo.sykmeldingsopplysninger.SykmeldingsOpplysningerClient
+import no.nav.syfo.sykmeldingsopplysninger.GetSykmeldingerDatabaseDevelopment
 import no.nav.syfo.utils.EnvironmentVariables
 import org.koin.core.KoinApplication
 import org.koin.dsl.module
@@ -138,7 +138,7 @@ val developmentSykmeldingModule = module {
 val developmentDokarkivModule = module { single<DokArkivClient> { DokarkivClientDevelopment() } }
 
 val developmentSykmeldingsopplysningerModule = module {
-    single<SykmeldingsOpplysningerClient> { DevelopmentSykmeldingsOpplysningerClient() }
+    single<GetSykmeldingOpplysningerDatabase> { GetSykmeldingerDatabaseDevelopment() }
     single {
         GetSykmeldingOpplysningerService(
             getSykmeldingOpplysningerDatabase = get(),
